Extracting EncodeBulkImportToProtobuf
