+++ packages/client/src/index.ts
@@ -5,3 +5,4 @@ export * from "./HistoryApi.js"
 export * from "./InspectionApi.js"
 export * from "./LanguagesApi.js"
 export * from "./RepositoryClient.js"
+export * from "./EncodeBulkImportToProtobuf.js"
